stage 6 mimic domain name from buff

+++ app/main.js
@@ -38,7 +38,7 @@ function getDomainName(buff){
   let index = 12
   let domain = ""
   while(true){
-  const length = buff.readInt8BE(index)
+  const length = buff.readInt16BE(index)
   if(length==0) break;
     domain += buff.toString("utf-8",index,index+length) + "."
     index += length
